requirements.txt
abstract-rendering==0.5.1
alabaster==0.7.7
anaconda-client==1.2.2
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
astropy==1.1.1
Babel==2.2.0
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.9.0
bokeh==0.11.0
boto==2.39.0
Bottleneck==1.0.0
cdecimal==2.3
cffi==1.2.1
clyent==1.2.0
colorama==0.3.6
configobj==5.0.6
cryptography==1.0.2
cycler==0.9.0
Cython==0.23.4
cytoolz==0.7.5
datashape==0.5.0
decorator==4.0.6
Django==1.9.2
django-bootstrap3==7.1.0
django-cors-headers==1.2.2
django-crispy-forms==1.6.1
django-extensions==1.7.4
docutils==0.12
dynd===f641248
ecdsa==0.13
enum34==1.1.2
et-xmlfile==1.0.1
Fabric==1.12.0
fastcache==1.0.2
Flask==0.10.1
funcsigs==0.4
futures==3.0.3
gevent==1.0.2
gevent-websocket==0.9.5
greenlet==0.4.9
grin==1.2.1
h5py==2.5.0
idna==2.0
ipaddress==1.0.14
ipdb==0.10.1
ipykernel==4.2.2
ipython==4.0.3
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jsonschema==2.4.0
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.1.0
jupyter-core==4.0.6
llvmlite==0.8.0
lxml==3.5.0
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.1
multipledispatch==0.4.8
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
nltk==3.1
nose==1.3.7
notebook==4.1.0
numba==0.23.1
numexpr==2.4.6
numpy==1.10.4
odo==0.4.0
openpyxl==2.3.2
pandas==0.17.1
paramiko==1.17.2
path.py==0.0.0
patsy==0.4.0
pep8==1.7.0
pexpect==3.3
pickleshare==0.5
Pillow==3.1.0
ply==3.8
psutil==3.4.2
psycopg2==2.6.2
ptyprocess==0.5
py==1.4.31
pyasn1==0.1.9
PyAudio==0.2.7
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.3
pyflakes==1.0.0
Pygments==2.1
pyOpenSSL==0.15.1
pyparsing==2.0.3
pytest==2.8.5
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
pyzmq==15.2.0
qtconsole==4.1.1
redis==2.10.3
requests==2.9.1
rope==0.9.4
scikit-image==0.11.3
scikit-learn==0.17
scipy==0.17.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.1
Sphinx==1.3.5
sphinx-rtd-theme==0.1.9
spyder==2.3.8
SQLAlchemy==1.0.11
statsmodels==0.6.1
sympy==0.7.6.1
tables==3.2.2
terminado==0.5
toolz==0.7.4
tornado==4.3
traitlets==4.1.0
unicodecsv==0.14.1
Werkzeug==0.11.3
xlrd==0.9.4
XlsxWriter==0.8.4
xlwings==0.6.4
xlwt==1.0.0
